backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
email-validator
requests
pydantic